Added comments, moved NoteAsyncHandler, updated gradle tools
