--- app/src/main/java/ru/dev2dev/notes/adapters/RecyclerViewCursorAdapter.java
@@ -1,39 +1,0 @@
-package ru.dev2dev.notes.adapters;
-
-import android.database.Cursor;
-import android.support.v7.widget.RecyclerView;
-
-public abstract class RecyclerViewCursorAdapter<VH extends RecyclerView.ViewHolder>
-        extends RecyclerView.Adapter<VH> {
-    private Cursor cursor;
-
-    public Cursor getCursor() {
-        return this.cursor;
-    }
-
-    public void swapCursor(final Cursor cursor) {
-        this.cursor = cursor;
-        notifyDataSetChanged();
-    }
-
-    public Cursor getItem(final int position) {
-        if (this.cursor != null && !this.cursor.isClosed()) {
-            this.cursor.moveToPosition(position);
-        }
-
-        return this.cursor;
-    }
-
-    @Override
-    public int getItemCount() {
-        return this.cursor != null ? this.cursor.getCount() : 0;
-    }
-
-    @Override
-    public final void onBindViewHolder(final VH holder, final int position) {
-        final Cursor cursor = getItem(position);
-        onBindViewHolder(holder, cursor);
-    }
-
-    public abstract void onBindViewHolder(final VH holder, final Cursor cursor);
-}
